refactor: update import paths for full transaction mall deferred pages
